+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Added
 - Added `2d_outer_only` configuration option.
+- Support 2-character problem navigation (aa-bz).